+++ lib/Service/OpmlService.php
@@ -15,33 +15,18 @@
 namespace OCA\News\Service;
 
 use OCA\News\Utility\OPMLExporter;
+use OCA\News\Utility\OPMLImporter;
+use OCA\News\Db\Folder;
 
 class OpmlService
 {
-
-    /**
-     * @var FolderServiceV2
-     */
-    private $folderService;
-
-    /**
-     * @var FeedServiceV2
-     */
-    private $feedService;
-
-    /**
-     * @var OPMLExporter
-     */
-    private $exporter;
-
     public function __construct(
-        FolderServiceV2 $folderService,
-        FeedServiceV2 $feedService,
-        OPMLExporter $exporter
+        private FolderServiceV2 $folderService,
+        private FeedServiceV2 $feedService,
+        private OPMLExporter $exporter,
+        private OPMLImporter $importer,
     ) {
-        $this->folderService = $folderService;
-        $this->feedService = $feedService;
-        $this->exporter = $exporter;
+        //NO-OP
     }
 
     /**
@@ -58,5 +43,49 @@ class OpmlService
 
         return $this->exporter->build($folders, $feeds)
                               ->saveXML();
+    }
+
+    /**
+     * Import all feeds and folders for a user.
+     *
+     * @param string $userId User ID
+     * @param string $data OPML data
+     *
+     * @return bool Status of the import
+     */
+    public function import(string $userId, string $data): bool
+    {
+        list($folders, $feeds) = $this->importer->import($userId, $data);
+
+        $folderEntities  = [];
+        $dbFolders = $this->folderService->findAllForUser($userId);
+        foreach ($folders as $folder) {
+            $existing = array_filter($dbFolders, fn(Folder $dbFolder) => $dbFolder->getName() === $folder['name']);
+
+            if (count($existing) > 0) {
+                $folderEntities[$folder['name']] = $existing[0];
+                continue;
+            }
+
+            $folderEntities[$folder['name']] = $this->folderService->create(
+                $userId,
+                $folder['name'],
+                $folderEntities[$folder['parentName']] ?? null,
+            );
+        }
+
+        foreach ($feeds as $feed) {
+            $parent = $folderEntities[$feed['folder']] ?? null;
+            $this->feedService->create(
+                $userId,
+                $feed['url'],
+                $parent?->getId(),
+                full_text: false,
+                title: $feed['title'],
+                full_discover: false,
+            );
+        }
+
+        return true;
     }
 }